requirements.txt
fastapi
uvicorn
pymongo[srv]
PyJWT
cryptography
passlib
bcrypt
python-multipart
httpx
python-dotenv
Pillow
slowapi
